feat: Adjust and fix rest of the UI related components

+++ features/merchant/shop_score/src/main/java/com/tokopedia/shop/score/penalty/presentation/model/calculation/ItemPenaltyCalculationScoreUiModel.kt
@@ -7,7 +7,7 @@ data class ItemPenaltyCalculationScoreUiModel(
     val score: Int,
     val date: String,
     val ongoingPoints: Int,
-    val totalOrder: Int,
+    val totalOrder: Long,
     val shopLevel: Int
 ): Visitable<PenaltyCalculationTypeFactory> {
